floats to double

+++ neuromlLocal/c302ForW2D.cpp
@@ -14,7 +14,7 @@ simulation(std::make_shared<SignalSimulatorForWorm2D>(simFileName,defaultSimClas
 //c302NervousSystem::c302NervousSystem():
 //simulation(new SignalSimulatorForWorm2D("neuromlLocal.main_sim",defaultSimClassName,0.005)){}
 
-c302ForW2D::c302ForW2D(float timestep_):
+c302ForW2D::c302ForW2D(double timestep_):
 inc_neuron_input_func("inc_neuron_input_j"),
 set_neuron_input_func("set_neuron_input_j"), v_newstep(new bool(true)), d_newstep(new bool(true)), 
 get_states_func("get_states_j"), get_output_func("get_outputs_j"),
