+++ neuromlLocal/owSignalSimulator.h
@@ -54,7 +54,7 @@
 
 class SignalSimulator: public owINeuronSimulator{
 public:
-  SignalSimulator(const std::string & simFileName, const std::string & simClassName, float timeStep);
+  SignalSimulator(const std::string & simFileName, const std::string & simClassName, double timeStep);
   //std::vector<float> run();
   void run();
   virtual ~SignalSimulator();
